fix modal

+++ src/components/Organization/organization-tabpane.tsx
@@ -5,9 +5,8 @@ import { Administrator, Company } from 'src/api/types';
 import { Util } from 'src/helpers/util';
 import { useAppDispatch } from 'src/redux/hooks';
 import { refreshCompanies } from 'src/redux/slices/companies/companies.slice';
+import { confirmation } from '../Modals/ConfirmationModal.component';
 import { OrganizationTable } from '../Table/organization-table';
-import { Modal } from 'antd';
-
 
 interface IOrganizationTabPane {
   trigger?: string;
@@ -16,7 +15,6 @@ interface IOrganizationTabPane {
 export const OrganizationTabPane = (props: IOrganizationTabPane) => {
   const { trigger } = props;
   const [loading, setLoading] = useState(false);
-  const [isModalVisible, setIsModalVisible] = useState(false);
 
   const [{ data, meta }, setData] = useState({
     data: [] as Administrator[],
@@ -39,26 +37,30 @@ export const OrganizationTabPane = (props: IOrganizationTabPane) => {
   }, [setLoading, setData, query]);
 
   const deleteOrganization = async (id: string) => {
-    showModal()
-    const toast = (await import('react-toastify')).toast;
     if (!loading) {
-      setLoading(true);
-      const clone = data.map((d) => ({ ...d }));
-      try {
-        setData({
-          meta,
-          data: data.filter((d) => (d.company as Company).id !== id),
-        });
-        await $api.company.deleteCompany(id);
-        getOrganizations();
-        refreshCompanies(dispatch);
-        toast.success('company deleted successfully');
-      } catch (error) {
-        const err = error as HttpError;
-        toast.error(err.message);
-        setData({ meta, data: clone });
-      } finally {
-        setLoading(false);
+      const shouldDelete = await confirmation({
+        text: 'Are you sure you want to permanently delete this organisation?',
+      });
+      if (shouldDelete) {
+        const toast = (await import('react-toastify')).toast;
+        setLoading(true);
+        const clone = data.map((d) => ({ ...d }));
+        try {
+          setData({
+            meta,
+            data: data.filter((d) => (d.company as Company).id !== id),
+          });
+          await $api.company.deleteCompany(id);
+          getOrganizations();
+          refreshCompanies(dispatch);
+          toast.success('company deleted successfully');
+        } catch (error) {
+          const err = error as HttpError;
+          toast.error(err.message);
+          setData({ meta, data: clone });
+        } finally {
+          setLoading(false);
+        }
       }
     }
   };
@@ -66,18 +68,6 @@ export const OrganizationTabPane = (props: IOrganizationTabPane) => {
   useEffect(() => {
     getOrganizations();
   }, [getOrganizations, trigger]);
-
-  const showModal = () => {
-    setIsModalVisible(true);
-  };
-
-  const handleOk = () => {
-    setIsModalVisible(false);
-  };
-
-  const handleCancel = () => {
-    setIsModalVisible(false);
-  };
 
   return (
     <div className="organisation__table-section">
@@ -88,16 +78,6 @@ export const OrganizationTabPane = (props: IOrganizationTabPane) => {
         deleteOrganisation={deleteOrganization}
         loading={loading}
       />
-<Modal title="Warning" 
-visible={isModalVisible} 
-onOk={handleOk} 
-onCancel={handleCancel} 
-okText="Delete"
- className="modalStyle">
-        <p className="organization-text-modal">You are about to delete this organzation, click on delete to continue?</p>
-      </Modal>
     </div>
-
   );
 };
-
